+++ src/stage.rs
@@ -42,9 +42,9 @@ impl Stage {
 
 #[derive(Clone, RustcEncodable, RustcDecodable)]
 pub struct Platform {
-    pub x:            f64,
-    pub y:            f64,
-    pub w:            f64,
-    pub h:            f64,
+    pub x:            f32,
+    pub y:            f32,
+    pub w:            f32,
+    pub h:            f32,
     pub pass_through: bool,
 }
